==> notebooks/01_synthetic_data_generation.ipynb <==
# Synthetic data generation

import os
if os.getcwd().split('/')[-1] == 'notebooks':
    os.chdir(os.pardir)

import matplotlib.pyplot as plt
import numpy as np
np.random.seed(0)
import pandas as pd
from copulas.univariate import GaussianKDE, TruncatedGaussian, UniformUnivariate
from sdv.constraints import Inequality, ScalarRange, ScalarInequality
from sdv.tabular import GaussianCopula, CTGAN, CopulaGAN
import seaborn as sns

from src.plotting import set_paper_context, set_poster_context

PATH_PROC_DATA = os.path.join('data', 'processed')
clean_data = pd.read_csv(os.path.join(PATH_PROC_DATA, 'pDeltaT_clean.csv'))

n_tot_rel_1 = Inequality(low_column_name='psPDn_1 [W/m2]',
                         high_column_name='psPDtot_1 [W/m2]')
n_tot_rel_4 = Inequality(low_column_name='psPDn_4 [W/m2]',
                         high_column_name='psPDtot_4 [W/m2]')
pDeltaT_min = ScalarInequality(column_name='pDeltaT * 100 [°C]',
                               relation='>=',
                               value=clean_data['pDeltaT * 100 [°C]'].min().item())
table_metadata = {'fields': {'d [mm]': {'type': 'numerical', 'subtype': 'float'},
                             'f [GHz]': {'type': 'numerical', 'subtype': 'float'},
                             'pPDn [W/m2]': {'type': 'numerical', 'subtype': 'float'},
                             'pPDtot [W/m2]': {'type': 'numerical', 'subtype': 'float'},
                             'psPDn_1 [W/m2]': {'type': 'numerical', 'subtype': 'float'},
                             'psPDtot_1 [W/m2]': {'type': 'numerical', 'subtype': 'float'},
                             'psPDn_4 [W/m2]': {'type': 'numerical', 'subtype': 'float'},
                             'psPDtot_4 [W/m2]': {'type': 'numerical', 'subtype': 'float'},
                             'pDeltaT * 100 [°C]': {'type': 'numerical', 'subtype': 'float'}},
                  'constraints': [n_tot_rel_1, n_tot_rel_4, pDeltaT_min]}
field_distributions = {'d [mm]': GaussianKDE,
                       'f [GHz]': GaussianKDE,
                       'pPDn [W/m2]': GaussianKDE,
                       'pPDtot [W/m2]': GaussianKDE,
                       'psPDn_1 [W/m2]': GaussianKDE,
                       'psPDtot_1 [W/m2]': GaussianKDE,
                       'psPDn_4 [W/m2]': GaussianKDE,
                       'psPDtot_4 [W/m2]': GaussianKDE,
                       'pDeltaT * 100 [°C]': GaussianKDE}
default_distribution = TruncatedGaussian
learn_rounding_scheme = False
enforce_min_max_values = True

def syn_data_gen(save=False):
    model = GaussianCopula(table_metadata=table_metadata,
                           field_distributions=field_distributions,
                           default_distribution=default_distribution,
                           learn_rounding_scheme=learn_rounding_scheme,
                           enforce_min_max_values=enforce_min_max_values)
    model.fit(clean_data)
    syn_data = model.sample(num_rows=1000)
    if save:
        syn_data.to_csv(os.path.join(PATH_PROC_DATA, 'pDeltaT_synthetic.csv'),
                        index=False)
    return model, syn_data

force_gen = False
save = False

if force_gen:
    print('Generating synthetic data...')
    model, syn_data = syn_data_gen(save)
else:
    try:
        syn_data = pd.read_csv(os.path.join(PATH_PROC_DATA, 'pDeltaT_synthetic.csv'))
    except Exception as e:
        print(e, '\nGenerating synthetic data...')
        model, syn_data = syn_data_gen()

## Quality report

from sdmetrics.reports.single_table import QualityReport

report = QualityReport()
report.generate(real_data=clean_data.drop(columns='pDeltaT [°C]'),
                synthetic_data=syn_data,
                metadata=table_metadata)

### Column shapes

The similarity of a real column vs. a synthetic column in terms of the column shapes - the marginal distribution of that column - by using the Kolmogorov-Smirnov (KS) test statistics. To compute this statistic, marginal distributions over real and synthetic data are converted into corresponding cumulative distribution functions. The KS statistic quantifies a distance between the empirical distribution functions. Here, the quality score is reported as 1 - KS statistic so that a higher score means higher quality.

col_shapes = report.get_details(property_name='Column Shapes')
display(col_shapes)

with set_paper_context():
    cs = sns.color_palette('rocket', 2)
    ps = ['00', '01', '10', '11']
    xs = ['psPDn_4 [W/m2]',
          'psPDtot_4 [W/m2]',
          'psPDn_1 [W/m2]',
          'psPDtot_1 [W/m2]']
    ls = [r'$psPD_{\text{n}, 4}$',
          r'$psPD_{\text{tot}, 4}$',
          r'$psPD_{\text{n}, 1}$',
          r'$psPD_{\text{tot}, 1}$']

    fig, axs = plt.subplots(2, 2, sharex=False, sharey=False, figsize=(4.5, 4))
    for i, x in enumerate(xs):
        irow, icol = ps[i]
        irow, icol = int(irow), int(icol)

        axs[irow, icol] = sns.histplot(data=clean_data, x=x,
                                       color=cs[0],
                                       stat='density', kde=True,
                                       label='collected', ax=axs[irow, icol],
                                       line_kws={'ls': '-', 'lw': 2})
        axs[irow, icol] = sns.histplot(data=syn_data, x=x,
                                       color=cs[1],
                                       stat='density', kde=True,
                                       label='synthetic', ax=axs[irow, icol],
                                       line_kws={'ls': '--', 'lw': 2})
        axs[irow, icol].set_title(ls[i], size=12)
        axs[irow, icol].set(xlabel='', ylabel='')

    fig.supxlabel(r'power density [W/m$^2$]')
    fig.supylabel(r'probability density')
    fig.suptitle('$\\mathbf{(a)}$', x=0.08, y=0.92)
    # handles, labels = fig.gca().get_legend_handles_labels()
    # by_label = dict(zip(labels, handles))
    # fig.legend(by_label.values(), by_label.keys(),
    #            title='data',
    #            bbox_to_anchor=(1.25, 0.615))
    fig.tight_layout()
    sns.despine()

    # fig_name = os.path.join('figures', 'syn_ipd_dist.png')
    # fig.savefig(fig_name, dpi=200, bbox_inches='tight')

with set_paper_context():
    cs = sns.color_palette('rocket', 2)

    fig, axs = plt.subplots(2, 1, sharex=True, figsize=(4.5, 4))
    axs[0] = sns.histplot(data=clean_data, x='pDeltaT * 100 [°C]',
                          color=cs[0],
                          stat='density', kde=True, 
                          label='collected', ax=axs[0],
                          line_kws={'ls': '-', 'lw': 2})
    axs[0] = sns.histplot(data=syn_data, x='pDeltaT * 100 [°C]',
                          color=cs[1],
                          stat='density', kde=True,
                          label='synthetic', ax=axs[0],
                          line_kws={'ls': '--', 'lw': 2})
    axs[0].set_title('probability density function', size=12)
    axs[0].set(xlabel='', ylabel='',
               xticks=[0, 60, 120],
               xticklabels=[0, 0.6, 1.2],
               xlim=[0, 120],
               yticks=[0, 0.025, 0.05],
               yticklabels=[0, 0.025, 0.05],
               ylim=[0, 0.05])
    axs[1] = sns.kdeplot(data=clean_data, x='pDeltaT * 100 [°C]',
                         cumulative=True, color=cs[0], ls='-', lw=2,
                         label='collected', ax=axs[1])
    axs[1] = sns.kdeplot(data=syn_data, x='pDeltaT * 100 [°C]',
                         cumulative=True, color=cs[1], ls='--', lw=2,
                         label='synthetic', ax=axs[1])
    axs[1].set_title('cumulative distribution function', size=12)
    axs[1].set(xlabel='', ylabel='',
               xticks=[0, 60, 120],
               xticklabels=[0, 0.6, 1.2],
               xlim=[0, 120],
               yticks=[0, 0.5, 1],
               yticklabels=[0, 0.5, 1],
               ylim=[0, 1])

    fig.supxlabel(r'temperature rise [°C]')
    fig.supylabel(r'probability density')
    fig.suptitle('$\\mathbf{(b)}$', x=0.08, y=0.92)
    # handles, labels = fig.gca().get_legend_handles_labels()
    # by_label = dict(zip(labels, handles))
    fig.legend(# by_label.values(), by_label.keys(),
               title='data',
               bbox_to_anchor=(1.3, 0.65))
    fig.tight_layout()
    sns.despine()

    # fig_name = os.path.join('figures', 'syn_temp_dist.png')
    # fig.savefig(fig_name, dpi=200, bbox_inches='tight')

with set_poster_context():
    cs = sns.color_palette('rocket', 2)
    ps = ['00', '01', '10', '11']
    xs = ['psPDn_4 [W/m2]',
          'psPDtot_4 [W/m2]',
          'psPDtot_1 [W/m2]',
          'pDeltaT * 100 [°C]']
    ls = [r'$psPD_{\text{n, 4}}$ (W/m$^\text{2}$)',
          r'$psPD_{\text{tot, 4}}$ (W/m$^\text{2}$)',
          r'$psPD_{\text{tot, 1}}$ (W/m$^\text{2}$)',
          r'$\Delta \text{T}_\text{max}$ $\cdot$ 100 °C']

    fig, axs = plt.subplots(2, 2, sharex=False, sharey=False, figsize=(6.5, 5.5))
    for i, x in enumerate(xs):
        irow, icol = ps[i]
        irow, icol = int(irow), int(icol)

        axs[irow, icol] = sns.histplot(data=clean_data, x=x,
                                       color=cs[0],
                                       stat='density', kde=True,
                                       label='collected', ax=axs[irow, icol],
                                       line_kws={'ls': '-', 'lw': 2})
        axs[irow, icol] = sns.histplot(data=syn_data, x=x,
                                       color=cs[1],
                                       stat='density', kde=True,
                                       label='synthetic', ax=axs[irow, icol],
                                       line_kws={'ls': '--', 'lw': 2})
        axs[irow, icol].set_title(ls[i])
        axs[irow, icol].set(xlabel='', ylabel='')

    fig.supylabel(r'probability density')
    handles, labels = fig.gca().get_legend_handles_labels()
    by_label = dict(zip(labels, handles))
    fig.legend(by_label.values(), by_label.keys(),
               title='data',
               bbox_to_anchor=(1.3, 0.6))
    fig.tight_layout()
    sns.despine()

    # fig_name = os.path.join('figures', 'poster', 'syn_dist.png')
    # fig.savefig(fig_name, dpi=500, bbox_inches='tight')

with set_paper_context():
    cs = sns.color_palette('rocket', 5)

    fig, ax = plt.subplots(1, 1, figsize=(4.5, 4))
    ax = sns.barplot(col_shapes.loc[4:],
                     x='Column', y='Quality Score',
                     edgecolor=cs, facecolor='None',
                     errwidth=1.5, capsize=0.1, lw=2, ax=ax)
    for container in ax.containers:
        ax.bar_label(container, rotation=90, label_type='center', c='k')
    ax.set(xlabel='', ylabel='',
           yticks=[0, 0.50, 1],
           yticklabels=[0, 0.5, 1], 
           ylim=[0, 1]
          )
    ax.set_xticklabels([r'$psPD_{\text{n}, 1}$',
                        r'$psPD_{\text{tot}, 1}$',
                        r'$psPD_{\text{n}, 4}$',
                        r'$psPD_{\text{tot}, 4}$',
                        r'$\Delta T_\text{max}$'], fontsize=11)
    fig.supylabel('similarity score')
    fig.suptitle('$\\mathbf{(c)}$', x=0.08, y=0.92)
    fig.tight_layout()
    sns.despine()

    # fig_name = os.path.join('figures', 'similarity_score.png')
    # fig.savefig(fig_name, dpi=200, bbox_inches='tight')

### Column pair trends

For a pair of columns, the column pair trend is computed as a Pearson correlation coefficient (assuming linear dependence) separateljy on real and synthetic data.
The correlation similarity reported as quality score is computed as the normalized relative difference between two separate correlation coefficients.

from scipy import stats

col_pair_trends = report.get_details(property_name='Column Pair Trends')
pDeltaT_pair_trend = col_pair_trends[
    col_pair_trends['Column 1'] == 'pDeltaT * 100 [°C]'
].reset_index(drop=True)
display(pDeltaT_pair_trend)

report.get_visualization(property_name='Column Pair Trends')

def _z_transform(r):
    return 0.5 * (np.log(1 + r) - np.log(1 - r))


def p_value(row, n1, n2):  # two-tailed
    z1 = _z_transform(row['Real Correlation'])
    z2 = _z_transform(row['Synthetic Correlation'])
    dz = z1 - z2
    z = dz / np.sqrt(1 / (n1 - 3) + 1 / (n2 - 3))
    p = 2 * stats.norm.sf(np.abs(z))
    if p < 0.05:
        rejected = True
    else:
        rejected = False
    return z, p, rejected


pDeltaT_pair_trend[['Z score', 'p value', 'H0 rejected']] = pDeltaT_pair_trend.apply(
    p_value, args=(100, 1000), axis=1, result_type='expand'
)
display(pDeltaT_pair_trend)

with set_paper_context():
    cs = sns.color_palette('rocket', 4)

    pDeltaT_pair_trend = pDeltaT_pair_trend.round({'Quality Score': 3})
    fig, axs = plt.subplots(2, 1, sharex=True, figsize=(4.5, 4))
    axs[0] = sns.pointplot(pDeltaT_pair_trend.loc[2:],
                           x='Column 2',
                           y='Real Correlation',
                           color='gray',
                           label='collected',
                           ax=axs[0])
    axs[0] = sns.pointplot(pDeltaT_pair_trend.loc[2:],
                           x='Column 2',
                           y='Synthetic Correlation',
                           color='darkgray',
                           markers='^',
                           linestyles='--',
                           label='synthetic',
                           ax=axs[0])
    axs[0].set(xlabel='', ylabel='corr. $\\Delta T_\\text{max}$',
               yticks=[0.7, 0.85, 1],
               yticklabels=[0.7, 0.85, 1],
               ylim=[0.7, 1]
              )

    axs[1] = sns.barplot(pDeltaT_pair_trend.loc[2:],
                         x='Column 2',
                         y='Quality Score',
                         edgecolor=cs, facecolor='None',
                         lw=2, ax=axs[1])
    for container in axs[1].containers:
        axs[1].bar_label(container, rotation=90, label_type='center', c='k')
    axs[1].set(xlabel='', ylabel='similarity score',
               yticks=[0, 0.50, 1],
               yticklabels=[0, 0.5, 1],
               xticklabels=[r'$psPD_{\text{n}, 1}$',
                            r'$psPD_{\text{tot}, 1}$',
                            r'$psPD_{\text{n}, 4}$',
                            r'$psPD_{\text{tot}, 4}$'], 
               ylim=[0, 1]
              )

    fig.legend(title='data', bbox_to_anchor=(1.3, 0.8))
    fig.suptitle('$\\mathbf{(d)}$', x=0.08, y=0.92)
    fig.tight_layout()
    sns.despine()

# fig_name = os.path.join('figures', 'column_pair.png')
# fig.savefig(fig_name, dpi=200, bbox_inches='tight')

with set_poster_context(font_scale=2):
    cs = sns.color_palette('rocket', 4)

    pDeltaT_pair_trend = pDeltaT_pair_trend.round({'Quality Score': 3})
    fig, axs = plt.subplots(2, 1, sharex=True, figsize=(6.5, 5.5))
    axs[0] = sns.pointplot(pDeltaT_pair_trend.loc[2:],
                           x='Column 2',
                           y='Real Correlation',
                           color='#4cc4b4',
                           label='collected',
                           ax=axs[0])
    axs[0] = sns.pointplot(pDeltaT_pair_trend.loc[2:],
                           x='Column 2',
                           y='Synthetic Correlation',
                           color='#4965a8',
                           markers='^',
                           linestyles='--',
                           label='synthetic',
                           ax=axs[0])
    axs[0].set(xlabel='', ylabel='corr. $\\Delta \\text{T}_\\text{max}$',
               yticks=[0.7, 1],
               yticklabels=[0.7, 1],
               ylim=[0.7, 1]
              )

    axs[1] = sns.barplot(pDeltaT_pair_trend.loc[2:],
                         x='Column 2',
                         y='Quality Score',
                         edgecolor=cs, facecolor='None',
                         lw=2, ax=axs[1])
    for container in axs[1].containers:
        axs[1].bar_label(container, rotation=90, label_type='center', c='k')
    axs[1].set(xlabel='power density', ylabel='sim. score',
               yticks=[0, 1],
               yticklabels=[0, 1],
               xticklabels=['n, 1',
                            'tot, 1',
                            'n, 4',
                            'tot, 4'], 
               ylim=[0, 1])

    fig.legend(ncols=2, bbox_to_anchor=(0.95, 1))
    fig.tight_layout()
    sns.despine()

    # fig_name = os.path.join('figures', 'poster', 'column_pair.png')
    # fig.savefig(fig_name, dpi=500, bbox_inches='tight')

## Regression tasks

Regression metrics calculate the success of using synthetic data to perform an ML regression task by using linear regression and multi-layer perceptron neural network. It first trains the ML algorithm using the training data (usually synthetic data). The output is an ML model that can predict the value of a given target column. Then, it tests the ML model by making predictions on the testing data (usually real data) and compares against the actual values. It finally returns the Train-Synthetic-Test-Real (TSTR) score by means of the coefficient of determination (R2).

from sdmetrics.single_table import LinearRegression

LinearRegression.compute(test_data=clean_data.drop(columns='pDeltaT [°C]'),
                         train_data=syn_data,
                         target='pDeltaT * 100 [°C]',
                         metadata=table_metadata)

## Diagnostic report

Few basic checks on the synthetic data to give a general sense of the strengths and weakness of the synthetic data generation  basic checks on your synthetic data to give a general sense of the strengths and weakness of your synthetic data model.  basic checks on your synthetic data to give a general sense of the strengths and weakness of your synthetic data model.  basic checks on your synthetic data to give a general sense of the strengths and weakness of your synthetic data model. model. 

from sdmetrics.reports.single_table import DiagnosticReport

report = DiagnosticReport()
report.generate(real_data=clean_data.drop(columns='pDeltaT [°C]'),
                synthetic_data=syn_data,
                metadata=table_metadata)

report.get_properties()

# Does the synthetic data cover the range of possible values?
report.get_details(property_name='Coverage')

# Is the synthetic data unique or does it copy the real rows?
report.get_details(property_name='Synthesis')

# Does the synthetic data respect the boundaries set by the real data?
report.get_details(property_name='Boundaries')  # score is the percentage of points within boundaries
