==> notebooks/03_multilayer_perceptron.ipynb <==
# Multilayer perceptron

import os
if os.getcwd().split('/')[-1] == 'notebooks':
    os.chdir(os.pardir)

import deepxde as dde
import matplotlib.pyplot as plt
import numpy as np
np.random.seed(0)
import pandas as pd

import seaborn as sns
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import mean_squared_error
from timeit import default_timer as timer

## Data

PATH_PROC_DATA = os.path.join('data', 'processed')
syn_data = pd.read_csv(os.path.join(PATH_PROC_DATA, 'pDeltaT_synthetic.csv'))
true_data = pd.read_csv(os.path.join(PATH_PROC_DATA, 'pDeltaT_clean.csv'))

features = ['d [mm]', 'f [GHz]', 'psPDtot_1 [W/m2]', 'psPDtot_4 [W/m2]']
target = 'pDeltaT * 100 [°C]'
X = syn_data[features]
y = syn_data[target] / 100

scaler = StandardScaler()
X_scaled = scaler.fit_transform(X)

# train and validation set - synthetic data
X_train, X_valid, y_train, y_valid = train_test_split(X_scaled,
                                                      y.to_numpy()[:, np.newaxis],
                                                      test_size=.2)
# test set - original data
X_test = true_data[features]
y_test = true_data[target] / 100
X_test_scaled = scaler.transform(X_test)

## Training the neural network

data = dde.data.DataSet(X_train=X_train,
                        X_test=X_valid,
                        y_train=y_train,
                        y_test=y_valid)

layer_size = [4] + [256] * 3 + [1]
activation = 'relu'
initializer = 'Glorot normal'
net = dde.nn.FNN(layer_size, activation, initializer)

model = dde.Model(data, net)

optimizer = 'adam'
learning_rate = 1e-3
model.compile(optimizer=optimizer,
              lr=learning_rate,
              metrics=['l2 relative error'])

def train(save=False):
    iterations = 100_000
    batch_size = 128
    display_every = int(iterations/10)
    start = timer()
    loss_history, train_state = model.train(iterations=iterations,
                                            batch_size=batch_size,
                                            display_every=display_every)
    end = timer()
    elapsed = start - end
    print(f'Training finished in {elapsed:.2f} s.')
    save_path = model.save(os.path.join('models', '03_multilayer_perceptron'))
    print(f'Model saved ({save_path}).')
    return model, loss_history

force_train = False
save = False

if force_train:
    print('Training the neural network...')
    model, loss_history = train(save)
    plot_loss = True
else:
    try:
        print('Trying to restore the neural network...')
        model.restore(os.path.join('models', '03_multilayer_perceptron-500000.pt'))
        print('Restoring successful.')
        plot_loss = False
    except Exception as e:
        print(e)
        print('Restoring failed. Training the neural network...')
        model, loss_history = train(save)
        plot_loss = True

if plot_loss:
    fig, ax = plt.subplots(1, 1, figsize=(4.5, 4))
    ax.plot(loss_history.loss_train, '-', label='train loss')
    ax.plot(loss_history.loss_test, '--', label='test loss')
    ax.plot(loss_history.metrics_test, ':', label='test metrics')
    ax.legend()
    ax.set(xlabel='iterations', ylabel='', yscale='log',)

## Evaluating the neural network

y_pred = model.predict(X_test_scaled)
rmse = np.sqrt(mean_squared_error(y_test, y_pred.ravel()))

y_resid = (y_test - y_pred.ravel())
ae = np.abs(y_resid)
mae = np.mean(ae)
mae

fig, ax = plt.subplots(1, 1, figsize=(4.5, 4))
ax = sns.histplot(x=ae, stat='density', kde=True, ax=ax)
ax.vlines(mae, *ax.get_ybound(), ls='--', label='mean absolute error')
ax.set(xlabel='absolute error (°C)', ylabel='probability density')
ax.legend();

PATH_ERROR_DATA = os.path.join('data', 'models')
error_data = os.path.join(PATH_ERROR_DATA, 'multilayer_perceptron.npy')
save = False
if save:
    with open(error_data, 'wb') as f:
        np.save(f, ae)
